Refactor main into TestSystem
